refactor: 인터페이스명 및 구현체명 변경
