+++ backend/src/main/java/com/woowacourse/gongcheck/application/AlertService.java
@@ -1,0 +1,8 @@
+package com.woowacourse.gongcheck.application;
+
+import com.woowacourse.gongcheck.application.response.SubmissionResponse;
+
+public interface AlertService {
+
+    void sendMessage(final SubmissionResponse submissionResponse);
+}
