+++ backend/src/test/java/com/woowacourse/gongcheck/documentation/DocumentationTest.java
@@ -6,7 +6,7 @@ import static org.springframework.restdocs.operation.preprocess.Preprocessors.pr
 import com.woowacourse.gongcheck.application.GuestAuthService;
 import com.woowacourse.gongcheck.application.JjwtTokenProvider;
 import com.woowacourse.gongcheck.application.JobService;
-import com.woowacourse.gongcheck.application.SlackService;
+import com.woowacourse.gongcheck.application.AlertService;
 import com.woowacourse.gongcheck.application.SpaceService;
 import com.woowacourse.gongcheck.application.SubmissionService;
 import com.woowacourse.gongcheck.application.TaskService;
@@ -55,7 +55,7 @@ class DocumentationTest {
     protected SubmissionService submissionService;
 
     @MockBean
-    protected SlackService slackService;
+    protected AlertService alertService;
 
     @MockBean
     protected JjwtTokenProvider jwtTokenProvider;
